implement main logic: rider can request a ride driver auto accept mode fee calculation trip complete, cancel

+++ gogo-backend/internal/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	ID    int    `db:"id" json:"id"`
+	ID    int64  `db:"id" json:"id"`
 	Name  string `db:"name" json:"name"`
 	Email string `db:"email" json:"email"`
 	Role  string `db:"role" json:"role"`
@@ -13,7 +13,7 @@ type LoginRequest struct {
 }
 
 type Register struct {
-	ID       int    `db:"id" json:"id"`
+	ID       int64  `db:"id" json:"id"`
 	Name     string `db:"name" json:"name"`
 	Email    string `db:"email" json:"email"`
 	Password string `db:"password" json:"password"`
