+++ gogo-backend/internal/model/vehicle.go
@@ -1,7 +1,7 @@
 package model
 
 type Vehicle struct {
-	ID          int    `db:"id" json:"id"`
+	ID          int64  `db:"id" json:"id"`
 	DriverID    int64  `db:"driver_id" json:"driver_id"`
 	VehicleType string `db:"vehicle_type" json:"vehicle_type"`
 	Make        string `db:"make" json:"make"`
